requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via django
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
django==3.0.9
    # via
    #   -r requirements.in
    #   djangorestframework
    #   drf-yasg
djangorestframework==3.12.1
    # via
    #   -r requirements.in
    #   drf-yasg
drf-yasg==1.20.0
    # via -r requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
markupsafe==2.1.0
    # via jinja2
packaging==21.3
    # via drf-yasg
psycopg2==2.8.6
    # via -r requirements.in
pyparsing==3.0.7
    # via packaging
pytz==2021.3
    # via django
requests==2.27.1
    # via coreapi
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.8
    # via requests
